// sniperbot/src/types.ts
export interface MintsDataReponse {
  solMint: string;
  tokenMint: string;
  baseVault?: string;
  quoteVault?: string;
  poolAddress?: string;
}



export interface QuoteResponse {
  data: unknown;
}

export interface SerializedQuoteResponse {
  swapTransaction: string;
  lastValidBlockHeight: number;
  prioritizationFeeLamports: number;
  computeUnitLimit: number;
  prioritizationType: {
    computeBudget: Record<string, unknown>;
  };
  simulationSlot: number;
  dynamicSlippageReport: {
    slippageBps: number;
    otherAmount: number;
    simulatedIncurredSlippageBps: number;
    amplificationRatio: string;
    categoryName: string;
    heuristicMaxSlippageBps: number;
  };
  simulationError: string | null;
}
// ...
export interface WebSocketRequest {
  jsonrpc: string;
  id: number;
  method: string;
  params: unknown[];
}
// ...
export interface CurrentTransaction {
  token: string;
  tokenTime: number;
  purchasePrice: number;
  balance: number;
  solPaid: number;
  solPaidUSDC: number;
  baseVault?: string;
  quoteVault?: string;
  poolAddress?: string;
}
// ...
export interface createSellTransactionResponse {
  success: boolean;
  msg: string | null;
  tx: string | null;
}